@@ -1,8 +1,8 @@
 {
   "_from": "mysql2",
-  "_id": "mysql2@2.3.0",
+  "_id": "mysql2@2.3.2",
   "_inBundle": false,
-  "_integrity": "sha512-0t5Ivps5Tdy5YHk5NdKwQhe/4Qyn2pload+S+UooDBvsqngtzujG1BaTWBihQLfeKO3t3122/GtusBtmHEHqww==",
+  "_integrity": "sha512-JUSA50rt/nSew8aq8xe3pRk5Q4y/M5QdSJn7Ey3ndOlPp2KXuialQ0sS35DNhPT5Z5PnOiIwSSQvKkl1WorqRA==",
   "_location": "/mysql2",
   "_phantomChildren": {},
   "_requested": {
@@ -19,8 +19,8 @@
     "#USER",
     "/"
   ],
-  "_resolved": "https://registry.npmjs.org/mysql2/-/mysql2-2.3.0.tgz",
-  "_shasum": "600f5cc27e397dfb77b59eac93666434f88e8079",
+  "_resolved": "https://registry.npmjs.org/mysql2/-/mysql2-2.3.2.tgz",
+  "_shasum": "3efe9814dbf1c2a3d7c2a1fc4666235939943ff9",
   "_spec": "mysql2",
   "_where": "C:\\Users\\Usuario\\Documents\\FATEC\\Fernando Martis 5º\\JPMuscle\\jpmuscle\\back",
   "author": {
@@ -32,9 +32,9 @@
   },
   "bundleDependencies": false,
   "dependencies": {
-    "denque": "^1.4.1",
+    "denque": "^2.0.1",
     "generate-function": "^2.3.1",
-    "iconv-lite": "^0.6.2",
+    "iconv-lite": "^0.6.3",
     "long": "^4.0.0",
     "lru-cache": "^6.0.0",
     "named-placeholders": "^1.1.2",
@@ -44,22 +44,22 @@
   "deprecated": false,
   "description": "fast mysql driver. Implements core protocol, prepared statements, ssl and compression in native JS",
   "devDependencies": {
-    "@typescript-eslint/eslint-plugin": "^2.3.2",
-    "@typescript-eslint/parser": "^2.3.2",
-    "assert-diff": "^3.0.0",
+    "@typescript-eslint/eslint-plugin": "^4.33.0",
+    "@typescript-eslint/parser": "^4.33.0",
+    "assert-diff": "^3.0.2",
+    "benchmark": "^2.1.4",
     "error-stack-parser": "^2.0.3",
-    "eslint": "^6.2.2",
-    "eslint-config-prettier": "^6.1.0",
+    "eslint": "^7.32.0",
+    "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-async-await": "0.0.0",
-    "eslint-plugin-markdown": "^1.0.0",
-    "husky": "^3.0.4",
+    "eslint-plugin-markdown": "^2.2.1",
+    "husky": "^7.0.2",
     "is-async-supported": "^1.2.0",
-    "lint-staged": "^10.0.3",
+    "lint-staged": "^11.2.0",
     "portfinder": "^1.0.28",
-    "prettier": "^2.0.5",
-    "prettier-markdown": "^0.1.6",
+    "prettier": "^2.4.1",
     "progress": "^2.0.3",
-    "typescript": "^3.6.3",
+    "typescript": "^4.4.3",
     "urun": "0.0.8",
     "utest": "0.0.8"
   },
@@ -102,17 +102,16 @@
     "url": "git+https://github.com/sidorares/node-mysql2.git"
   },
   "scripts": {
-    "benchmark": "./benchmarks/run-unit.js",
+    "benchmark": "node ./benchmarks/benchmark.js",
     "eslint-check": "eslint --print-config .eslintrc | eslint-config-prettier-check",
     "lint": "npm run lint:docs && npm run lint:code",
     "lint:code": "eslint index.js promise.js \"lib/**/*.js\" \"test/**/*.js\" \"benchmarks/**/*.js\"",
     "lint:docs": "eslint Contributing.md \"documentation/**/*.md\" \"examples/*.js\"",
     "precommit": "lint-staged",
     "prettier": "prettier --single-quote --trailing-comma none --write \"{lib,examples,test}/**/*.js\"",
-    "prettier:docs": "prettier-markdown README.md documentation/*",
-    "test": "npm run lint && npm run test:raw",
-    "test:raw": "node ./test/run.js",
+    "prettier:docs": "prettier --single-quote --trailing-comma none --write README.md documentation/*",
+    "test": "node ./test/run.js",
     "wait-port": "wait-on"
   },
-  "version": "2.3.0"
+  "version": "2.3.2"
 }
