@@ -1,10 +1,11 @@
+const db = require('../../config/db.js');
+
 module.exports = {
-    indexControllers,
+    indexControllers
 }
 
 function indexControllers(req, res)  {
     console.log('Rota Raiz Encontrada!');
     res.json('Rota Raiz Encontrada!');
 };
-
 
