@@ -1,9 +1,10 @@
 const userModels = require('../models/userModels.js');
-
+const db = require('../../config/db')
 
 module.exports = {
   userMenu,
-  userGetAll
+  userGetAll,
+  userInsert
 }
 
 function userMenu(req, res) {
@@ -13,14 +14,28 @@ function userMenu(req, res) {
 
 function userGetAll(req, res) {
   console.log('Listar Users { M O D E L S }!');
-  //res.json('Listar Usuario { M O D E L S }!')
-  userModels.getAllUser(function (err, resposta) {
-    console.log(`Retorno de Users { M O D E L S }`);
 
+  const sqlGet = "SELECT * FROM  users"
+  db.query(sqlGet, (err, result) => {
     if (err) {
       throw err;
     } else {
-      res.json(resposta);
+      res.send(result);
     }
-  })
+  });
 }
+
+  function userInsert(req, res) {
+    console.log("Rota Insert User Encontrada!");
+
+    const nome = req.body.nomeUsuario;
+    const senha = req.body.emailUsuario;
+    const sqlInsert = "INSERT INTO users (nomeUsuario, senha)VALUES (?,?)";
+    db.query(sqlInsert, [nome, senha], (err, result) => {
+      if (err) {
+        throw err;
+      } else {
+        res.json(result);
+      }
+    })
+  }
