@@ -1,10 +1,6 @@
 const controllersCadastro = require('../controllers/cadastroControllers.js');
 
 
-server.get('/cadastro', controllersCadastro.cadastroMenu);
-
 server.get('/cadastro/listar', controllersCadastro.cadastroGetAll);
 
-server.post('/cadastro/adicionar', controllersCadastro.cadastroPost);
-
-
+server.post('/cadastro/insert', controllersCadastro.cadastroInsert);
