@@ -1,11 +1,14 @@
+/*
 const express = require("express");
 const consign = require("consign");
 const cors = require("cors")
 // Iniciando o Servidor Express
 server = express();
 server.use(cors());
+server.use(express.json())
 
-server.set('porta', 3001);
+//server.set('porta', 3001);
+//server.set('url', 'https://localhost:3001')
 
 consign({ cwd: 'api'})
     .include('models')
@@ -14,4 +17,8 @@ consign({ cwd: 'api'})
     .into(server)
 ;
 
+server.listen(port, function() {
+  
+})
 module.exports = server;
+*/
