@@ -2,21 +2,21 @@ const mysql = require('mysql2');
 const database = 'cliente';
 
 // instanciar objeto de acesso ao banco de dados
-const conexao = mysql.createConnection({
+const db = mysql.createConnection({
     user: 'root',
     password: '130517',
     host: 'localhost',
     port: 3306
 });
 
-conexao.connect((err) => {
+db.connect((err) => {
     if (err) {
         console.log('Erro ao conectar no mysql...', err)
         return
     }
-    conexao.query('USE ' + database);
+    db.query('USE ' + database);
     console.log('\nConexão estabilizada com sucesso!!!')
 })
 
-module.exports = conexao;
+module.exports = db;
 
